Swapped from Crud Repository to JPA Repository, Added Unit Test

+++ src/main/java/com/wsd/WeatherSensorData/entity/Sensor.java
@@ -58,7 +58,7 @@ public class Sensor
         this.city = city;
     }
     
-    public float getTemperature()
+    public int getTemperature()
     {
         return this.temperature;
     }
